requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-docx
PyMuPDF
openai
